+++ R/createQTLMPP.R
@@ -26,7 +26,7 @@
 #'                   evalDist = 5)
 #'
 #' ## Multi-QTL Mapping.
-#' ABC_MQM <- selQTLMPP(ABC, trait = "pheno")
+#' ABC_MQM <- selQTLMPP(ABC, trait = "yield")
 #'
 #' ## Print summary.
 #' summary(ABC_MQM)
@@ -181,7 +181,7 @@ print.summary.QTLMPP <- function(x,
 #'                   evalDist = 5)
 #'
 #' ## Multi-QTL Mapping.
-#' ABC_MQM <- selQTLMPP(ABC, trait = "pheno")
+#' ABC_MQM <- selQTLMPP(ABC, trait = "yield")
 #'
 #' ## QTL Profile plot.
 #' plot(ABC_MQM, plotType = "QTLProfile")