+++ R/randomQTLmodel.R
@@ -5,7 +5,7 @@
 randomQTLmodel <- function(modDat,
                            map,
                            parents,
-                           trait = "pheno",
+                           trait = "yield",
                            scanMrk = NULL,
                            cofMrk = NULL,
                            NULLmodel = FALSE) {