+++ R/selQTLMPP.R
@@ -8,8 +8,18 @@
 #' no new markers with a \eqn{-log10(p)} value above the threshold or until
 #' the maximum number of cofactors is reached.
 #'
+#' By default only family specific effects and residual variances and no
+#' kinship relations are included in the model. It is possible to include
+#' kinship relations by either specifying \code{computeKin = TRUE}. When doing
+#' so the kinship matrix is computed by averaging \eqn{Z Z^t} over all markers,
+#' where \eqn{Z} is the genotype x parents matrix for the marker. It is also
+#' possible to specify a list of precomputed chromosome
+#' specific kinship matrices in \code{K}. Note that adding a kinship matrix
+#' to the model increases the computation time a lot, especially for large
+#' populations.
+#'
 #' @param MPPobj An object of class gDataMPP, typically the output of either
-#' \code{\link{calcIBDMPP}} or \code{\link{readRABBIT}}.
+#' \code{\link{calcIBDMPP}} or \code{\link{readRABBITMPP}}.
 #' @param trait A character string indicating the trait QTL mapping is done for.
 #' @param QTLwindow A numerical value indicating the window around a QTL that
 #' is considered as part of that QTL.
@@ -18,6 +28,12 @@
 #' @param maxCofactors A numerical value, the maximum number of cofactors to
 #' include in the model. If \code{NULL} cofactors are added until no new
 #' cofactors are found.
+#' @param K A list of chromosome specific kinship matrices. If
+#' \code{NULL} and \code{computeKin = FALSE} no kinship matrix is included in
+#' the models.
+#' @param computeKin Should chromosome specific kinship matrices be computed?
+#' @param parallel Should the computation of variance components be done in
+#' parallel? This requires a parallel back-end to be registered. See examples.
 #' @param verbose Should progress and intermediate plots be output?
 #'
 #' @return An object of class \code{QTLMPP}
@@ -43,12 +59,22 @@
 #'                   evalDist = 5)
 #'
 #' ## Single-QTL Mapping.
-#' ABC_SQM <- selQTLMPP(ABC, trait = "pheno", maxCofactors = 0)
+#' ABC_SQM <- selQTLMPP(ABC, trait = "yield", maxCofactors = 0)
 #'
-#' ## Multi-QTL Mapping
+#' ## Multi-QTL Mapping.
 #' \dontrun{
-#' ABC_MQM <- selQTLMPP(ABC, trait = "pheno")
+#' ## Register parallel back-end with 2 cores.
+#' doParallel::registerDoParallel(cores = 2)
+#'
+#' ## Run multi-QTL mapping.
+#' ABC_MQM <- selQTLMPP(ABC, trait = "yield", parallel = TRUE)
+#'
+#' ## Run multi-QTL mapping - include kinship matrix.
+#' ABC_MQM_kin <- selQTLMPP(ABC, trait = "yield", parallel = TRUE,
+#'                         computeKin = TRUE)
 #' }
+#'
+#' @seealso \code{\link{kinshipIBD}}
 #'
 #' @importFrom utils head tail
 #' @export
@@ -57,6 +83,9 @@ selQTLMPP <- function(MPPobj,
                       QTLwindow = 10,
                       threshold = 3,
                       maxCofactors = NULL,
+                      K = NULL,
+                      computeKin = FALSE,
+                      parallel = FALSE,
                       verbose = FALSE) {
   if (!inherits(MPPobj, "gDataMPP")) {
     stop("MPPobj should be an object of class gDataMPP.\n")
@@ -93,25 +122,63 @@ selQTLMPP <- function(MPPobj,
        maxCofactors < 0)) {
     stop("maxCofactors should be a positive numerical value.\n")
   }
+  if (!is.null(K) &&
+      (!is.list(K) || !all(sapply(K, FUN = function(k) {
+        is.matrix(k) || inherits(k, "Matrix")})) ||
+       length(K) != length(unique(map$chr)))) {
+    stop("K should be a list of matrices of length equal to the ",
+         "number of chromosomes in the map.\n")
+  }
   if (is.null(maxCofactors)) {
     maxCofactors <- dim(markers)[2]
   }
   parents <- dimnames(markers)[[3]]
   nPar <- length(parents)
-  markerNames <- dimnames(markers)[[2]]
-  map <- MPPobj$map
+  markerNames <- colnames(markers)
+  ## Restrict map to markers in markers.
+  map <- map[rownames(map) %in% markerNames, ]
   ## Construct model data by merging phenotypic and genotypic data.
   ## Merge phenotypic data and covar (cross).
-  modDat <- merge(pheno, covar, by.x = "genotype", by.y = "row.names")
+  if (!is.null(covar) && hasName(x = covar, name = "cross")) {
+    modDat <- merge(pheno, covar, by.x = "genotype", by.y = "row.names")
+  } else {
+    modDat <- pheno
+    modDat[["cross"]] <- factor(1)
+  }
   ## Remove missing values for trait from modDat.
   ## Not strictly necessary, but it prevents warnings from LMMsolve later on.
   modDat <- droplevels(modDat[!is.na(modDat[[trait]]), ])
-  ## Flatten markers to 2D structure.
-  markers <- do.call(cbind, apply(X = markers, MARGIN = 2,
-                                  FUN = I, simplify = FALSE))
-  colnames(markers) <- paste0(rep(markerNames, each = nPar), "_", parents)
-  ## Merge markers to modDat.
-  modDat <- merge(modDat, markers, by.x = "genotype", by.y = "row.names")
+  ## Restrict markers to genotypes in modDat.
+  markers <- markers[rownames(markers) %in% modDat[["genotype"]], , ]
+  ## Restrict modDat to genotypes in markers.
+  modDat <- droplevels(modDat[modDat[["genotype"]] %in% rownames(markers), ])
+  ## Compute kinship matrices.
+  if (computeKin) {
+    K <- kinshipIBD(map = map, markers = markers)
+  } else if (!is.null(K)) {
+    K <- sapply(K, FUN = function(KChr) {
+        ## Remove genotypes not in data.
+        KChr[rownames(KChr) %in% modDat[["genotype"]],
+             rownames(KChr) %in% modDat[["genotype"]]]
+        ## Assure order of genotypes is identical to that in markers.
+        KChr[order(match(rownames(KChr), rownames(markers))),
+             order(match(colnames(KChr), rownames(markers)))]
+    }, simplify = FALSE)
+  } else {
+    K <- NULL
+  }
+  if (!is.null(K)) {
+    ## Compute spectral decomposition.
+    Usc <- sapply(K, FUN = function(KChr) {
+      eigKChr <- eigen(KChr)
+      selEigVals <- eigKChr$values > sqrt(.Machine$double.eps)
+      U <- eigKChr$vectors[, selEigVals]
+      d <- eigKChr$values[selEigVals]
+      U %*% diag(sqrt(d))
+    }, simplify = FALSE)
+  } else {
+    Usc <- NULL
+  }
   ## Initialize parameters.
   cofactors <- NULL
   mapScan <- map
@@ -122,10 +189,13 @@ selQTLMPP <- function(MPPobj,
     }
     scanRes <- scanQTL(modDat = modDat,
                        map = mapScan,
+                       markers = markers,
                        parents = parents,
                        trait = trait,
                        QTLwindow = QTLwindow,
                        cof = cofactors,
+                       Usc = Usc,
+                       parallel = parallel,
                        verbose = verbose)
     if (verbose) {
       plotIntermediateScan(scanRes,
@@ -144,7 +214,7 @@ selQTLMPP <- function(MPPobj,
     ## Add new cofactor to list of cofactors for next round of scanning.
     cofactors <- c(cofactors, scanSel[which.max(scanSel[["minlog10p"]]), "snp"])
     qtlWinScan <- scanRes[scanRes[["QTLRegion"]] &
-                             !scanRes[["snp"]] %in% cofactors, "snp"]
+                            !scanRes[["snp"]] %in% cofactors, "snp"]
     if (length(cofactors) == 1) {
       lowPScan <- scanRes[!is.na(scanRes[["minlog10p"]]) &
                             scanRes[["minlog10p"]] < 0.31, "snp"]
@@ -154,11 +224,23 @@ selQTLMPP <- function(MPPobj,
   ## Final run with all markers and all cofactors.
   scanRes <- scanQTL(modDat = modDat,
                      map = map,
+                     markers = markers,
                      parents = parents,
                      trait = trait,
                      QTLwindow = QTLwindow,
                      cof = cofactors,
+                     Usc = Usc,
+                     parallel = parallel,
                      verbose = verbose)
+  ## Flatten cofactor markers to 2D structure.
+  if (!is.null(cofactors)) {
+    markersCof <- do.call(cbind, lapply(X = cofactors, FUN = function(mrk) {
+      markers[, mrk, ]
+    }))
+    colnames(markersCof) <- paste0(rep(cofactors, each = nPar), "_", parents)
+    ## Merge markers to modDat.
+    modDat <- merge(modDat, markersCof, by.x = "genotype", by.y = "row.names")
+  }
   ## Fit model with all cofactors for computation of variance explained.
   finMod <- randomQTLmodel(modDat = modDat, map = map, parents = parents,
                            trait = trait, scanMrk = NULL, cofMrk = cofactors,
@@ -195,6 +277,7 @@ selQTLMPP <- function(MPPobj,
   GWASInfo <- list(parents = parents)
   res <- createGWAS(GWAResult = list(pheno = GWARes),
                     signSnp = list(pheno = signSnp),
+                    kin = K,
                     thr = list(pheno = setNames(threshold, trait)),
                     GWASInfo = GWASInfo)
   ## Add QTLMPP class to simplify providing generic functions.