+++ R/statgenMPP.R
@@ -2,4 +2,7 @@
 #'
 #' @keywords internal
 "_PACKAGE"
-NULL
+## CRAN Note avoidance.
+if (getRversion() >= "2.15.1") {
+  utils::globalVariables("i")
+}